Refactor TemplatePluginExtension to remove event handling; introduce TemplatePluginExtensionWithListener for event management
