+++ src/main/java/com/epam/reportportal/extension/template/TemplatePluginExtension.java
@@ -3,36 +3,26 @@ package com.epam.reportportal.extension.template;
 import com.epam.reportportal.extension.CommonPluginCommand;
 import com.epam.reportportal.extension.PluginCommand;
 import com.epam.reportportal.extension.ReportPortalExtensionPoint;
-import com.epam.reportportal.extension.event.PluginEvent;
 import com.epam.reportportal.extension.template.command.TemplateCommand;
-import com.epam.reportportal.extension.template.event.plugin.PluginEventHandlerFactory;
-import com.epam.reportportal.extension.template.event.plugin.PluginEventListener;
 import com.epam.reportportal.extension.template.utils.MemoizingSupplier;
-import com.epam.ta.reportportal.dao.IntegrationRepository;
-import com.epam.ta.reportportal.dao.IntegrationTypeRepository;
 import com.epam.ta.reportportal.dao.LogRepository;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.function.Supplier;
-import jakarta.annotation.PostConstruct;
 import org.pf4j.Extension;
-import org.springframework.beans.factory.DisposableBean;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.ApplicationListener;
-import org.springframework.context.event.ApplicationEventMulticaster;
-import org.springframework.context.support.AbstractApplicationContext;
 import org.springframework.stereotype.Service;
 
 /**
- * Use this extension if you need to handle events and create a default integration.
+ * Use this extension class if you don't need to handle any events and create a default
+ * integration. Moreover, you can delete java/com/epam/reportportal/extension/template/event.
  *
- * @author Andrei Piankouski
+ * @author Reingold Shekhtel
  */
 @Extension
 @Service
-public class TemplatePluginExtension implements ReportPortalExtensionPoint, DisposableBean {
+public class TemplatePluginExtension implements ReportPortalExtensionPoint {
 
   private final Supplier<Map<String, PluginCommand>> pluginCommandMapping = new MemoizingSupplier<>(
       this::getCommands);
@@ -42,50 +32,6 @@ public class TemplatePluginExtension implements ReportPortalExtensionPoint, Disp
 
   @Autowired
   private LogRepository logRepository;
-
-  @Autowired
-  private ApplicationContext applicationContext;
-
-  private final Supplier<ApplicationListener<PluginEvent>> pluginLoadedListener;
-
-  @Autowired
-  private IntegrationTypeRepository integrationTypeRepository;
-
-  @Autowired
-  private IntegrationRepository integrationRepository;
-
-  public TemplatePluginExtension() {
-    pluginLoadedListener = new MemoizingSupplier<>(() -> new PluginEventListener(
-        "template",
-        new PluginEventHandlerFactory(integrationTypeRepository, integrationRepository)
-    ));
-  }
-
-  @PostConstruct
-  public void createIntegration() {
-    initListeners();
-  }
-
-  private void initListeners() {
-    ApplicationEventMulticaster applicationEventMulticaster = applicationContext.getBean(
-        AbstractApplicationContext.APPLICATION_EVENT_MULTICASTER_BEAN_NAME,
-        ApplicationEventMulticaster.class
-    );
-    applicationEventMulticaster.addApplicationListener(pluginLoadedListener.get());
-  }
-
-  @Override
-  public void destroy() {
-    removeListeners();
-  }
-
-  private void removeListeners() {
-    ApplicationEventMulticaster applicationEventMulticaster = applicationContext.getBean(
-        AbstractApplicationContext.APPLICATION_EVENT_MULTICASTER_BEAN_NAME,
-        ApplicationEventMulticaster.class
-    );
-    applicationEventMulticaster.removeApplicationListener(pluginLoadedListener.get());
-  }
 
   @Override
   public Map<String, ?> getPluginParams() {
